Issue 42476: Support JSON and JSONB operators and functions on Postges
