+++ core/src/org/labkey/core/CoreModule.java
@@ -395,6 +395,8 @@ public class CoreModule extends SpringModule implements SearchService.DocumentPr
         registerHealthChecks();
 
         ContextListener.addNewInstallCompleteListener(() -> sendSystemReadyEmail(UserManager.getAppAdmins()));
+
+        PostgreSqlDialectFactory.registerJSONFunctions();
     }
 
     private void registerHealthChecks()
