+++ core/src/org/labkey/core/dialect/PostgreSqlDialectFactory.java
@@ -25,12 +25,18 @@ import org.junit.Assert;
 import org.junit.Test;
 import org.labkey.api.collections.CsvSet;
 import org.labkey.api.data.DbScope;
+import org.labkey.api.data.JdbcType;
+import org.labkey.api.data.SQLFragment;
 import org.labkey.api.data.dialect.AbstractDialectRetrievalTestCase;
 import org.labkey.api.data.dialect.DatabaseNotSupportedException;
 import org.labkey.api.data.dialect.JdbcHelperTest;
 import org.labkey.api.data.dialect.SqlDialect;
 import org.labkey.api.data.dialect.SqlDialectFactory;
+import org.labkey.api.data.dialect.SqlDialectManager;
 import org.labkey.api.data.dialect.TestUpgradeCode;
+import org.labkey.api.query.AbstractMethodInfo;
+import org.labkey.api.query.QueryParseException;
+import org.labkey.api.query.QueryService;
 import org.labkey.api.util.HtmlString;
 import org.labkey.api.util.PageFlowUtil;
 import org.labkey.api.util.VersionNumber;
@@ -54,6 +60,77 @@ public class PostgreSqlDialectFactory implements SqlDialectFactory
     {
         // PostgreSQL JDBC driver should not be present in <tomcat>/lib
         DbScope.registerForbiddenTomcatFilenamePredicate(filename->filename.equalsIgnoreCase("postgresql.jar"));
+    }
+
+    public static void registerJSONFunctions()
+    {
+        SqlDialect postgresDialect = SqlDialectManager.getFromDriverClassname(null, "org.postgresql.Driver");
+
+        // Wire up JSON and JSONB data type support for Postgres, as described here:
+        // https://www.postgresql.org/docs/9.5/functions-json.html
+
+        // Pretend that the JSON operators are a function instead so that we don't need them to be fully supported
+        // for query parsing
+
+        QueryService.get().registerMethod("json_op", new AbstractMethodInfo(JdbcType.VARCHAR)
+        {
+            private final Set<String> ALLOWED_OPERATORS = Set.of("->", "->>", "#>", "#>>", "@>", "<@",
+                    "?", "?|", "?&", "||", "-", "#-");
+
+            @Override
+            public SQLFragment getSQL(SqlDialect dialect, SQLFragment[] arguments)
+            {
+                SQLFragment rawOperator = arguments[1];
+                String operatorRawString = rawOperator.getSQL();
+                if (!rawOperator.getParams().isEmpty() || !operatorRawString.startsWith("'") || !operatorRawString.endsWith("'"))
+                {
+                    throw new QueryParseException("Unsupported JSON operator: " + rawOperator, null, 0, 0);
+                }
+
+                String strippedOperator = operatorRawString.substring(1, operatorRawString.length() - 1);
+                if (!ALLOWED_OPERATORS.contains(strippedOperator))
+                {
+                    throw new QueryParseException("Unsupported JSON operator: " + rawOperator, null, 0, 0);
+                }
+
+                return new SQLFragment("(").append(arguments[0]).append(")").
+                        append(strippedOperator).
+                        append("(").append(arguments[2]).append(")");
+            }
+        }, JdbcType.VARCHAR, 3, 3);
+
+        QueryService.get().registerPassthroughMethod("to_json", null, JdbcType.OTHER, 1, 1, postgresDialect);
+        QueryService.get().registerPassthroughMethod("to_jsonb", null, JdbcType.OTHER, 1, 1, postgresDialect);
+        QueryService.get().registerPassthroughMethod("array_to_json", null, JdbcType.OTHER, 1, 2, postgresDialect);
+        QueryService.get().registerPassthroughMethod("row_to_json", null, JdbcType.OTHER, 1, 2, postgresDialect);
+
+        addJsonPassthroughMethod("build_array", JdbcType.OTHER, 1, Integer.MAX_VALUE, postgresDialect);
+        addJsonPassthroughMethod("build_object", JdbcType.OTHER, 1, Integer.MAX_VALUE, postgresDialect);
+        addJsonPassthroughMethod("object", JdbcType.OTHER, 1, 2, postgresDialect);
+
+        addJsonPassthroughMethod("array_length", JdbcType.INTEGER, 1, 1, postgresDialect);
+        addJsonPassthroughMethod("each", JdbcType.OTHER, 1, 1, postgresDialect);
+        addJsonPassthroughMethod("each_text", JdbcType.OTHER, 1, 1, postgresDialect);
+        addJsonPassthroughMethod("extract_path", JdbcType.OTHER, 2, Integer.MAX_VALUE, postgresDialect);
+        addJsonPassthroughMethod("extract_path_text", JdbcType.VARCHAR, 2, Integer.MAX_VALUE, postgresDialect);
+        addJsonPassthroughMethod("object_keys", JdbcType.OTHER, 1, 1, postgresDialect);
+        addJsonPassthroughMethod("populate_record", JdbcType.OTHER, 2, 2, postgresDialect);
+        addJsonPassthroughMethod("populate_recordset", JdbcType.OTHER, 2, 2, postgresDialect);
+        addJsonPassthroughMethod("array_elements", JdbcType.OTHER, 1, 1, postgresDialect);
+        addJsonPassthroughMethod("array_elements_text", JdbcType.VARCHAR, 1, 1, postgresDialect);
+        addJsonPassthroughMethod("typeof", JdbcType.VARCHAR, 1, 1, postgresDialect);
+        addJsonPassthroughMethod("to_record", JdbcType.OTHER, 1, 1, postgresDialect);
+        addJsonPassthroughMethod("to_recordset", JdbcType.OTHER, 1, 1, postgresDialect);
+        addJsonPassthroughMethod("strip_nulls", JdbcType.OTHER, 1, 1, postgresDialect);
+
+        QueryService.get().registerPassthroughMethod("jsonb_set", null, JdbcType.OTHER, 3, 4, postgresDialect);
+        QueryService.get().registerPassthroughMethod("jsonb_pretty", null, JdbcType.VARCHAR, 1, 1, postgresDialect);
+    }
+
+    private static void addJsonPassthroughMethod(String name, JdbcType type, int minArgs, int maxArgs, SqlDialect dialect)
+    {
+        QueryService.get().registerPassthroughMethod("json_" + name, null, type, minArgs, maxArgs, dialect);
+        QueryService.get().registerPassthroughMethod("jsonb_" + name, null, type, minArgs, maxArgs, dialect);
     }
 
     @Override
@@ -119,9 +196,9 @@ public class PostgreSqlDialectFactory implements SqlDialectFactory
     }
 
     @Override
-    public Collection<? extends Class> getJUnitTests()
+    public Collection<? extends Class<?>> getJUnitTests()
     {
-        return Arrays.<Class>asList(DialectRetrievalTestCase.class, InlineProcedureTestCase.class, JdbcHelperTestCase.class);
+        return Arrays.asList(DialectRetrievalTestCase.class, InlineProcedureTestCase.class, JdbcHelperTestCase.class);
     }
 
     @Override
