+++ mothership/src/org/labkey/mothership/MothershipModule.java
@@ -82,81 +82,6 @@ public class MothershipModule extends DefaultModule
     {
         addController("mothership", MothershipController.class);
         MothershipSchema.register(this);
-
-        SqlDialect postgresDialect = SqlDialectManager.getFromDriverClassname(null, "org.postgresql.Driver");
-
-        // Wire up JSON and JSONB data type support for Postgres, as described here:
-        // https://www.postgresql.org/docs/9.5/functions-json.html
-        // Ultimately we need to promote these out of the mothership module but this will enable it for JSON metric
-        // reporting for now.
-
-        // Pretend that the JSON operators are a function instead so that we don't need them to be fully supported
-        // for query parsing
-
-        if (MothershipManager.get().getDialect().isPostgreSQL())
-        {
-            QueryService.get().registerMethod("json_op", new AbstractMethodInfo(JdbcType.VARCHAR)
-            {
-                private final Set<String> ALLOWED_OPERATORS = Set.of("->", "->>", "#>", "#>>", "@>", "<@",
-                        "?", "?|", "?&", "||", "-", "#-");
-
-                @Override
-                public SQLFragment getSQL(SqlDialect dialect, SQLFragment[] arguments)
-                {
-                    SQLFragment rawOperator = arguments[1];
-                    String operatorRawString = rawOperator.getSQL();
-                    if (!rawOperator.getParams().isEmpty() || !operatorRawString.startsWith("'") || !operatorRawString.endsWith("'"))
-                    {
-                        throw new QueryParseException("Unsupported JSON operator: " + rawOperator, null, 0, 0);
-                    }
-
-                    String strippedOperator = operatorRawString.substring(1, operatorRawString.length() - 1);
-                    if (!ALLOWED_OPERATORS.contains(strippedOperator))
-                    {
-                        throw new QueryParseException("Unsupported JSON operator: " + rawOperator, null, 0, 0);
-                    }
-
-                    return new SQLFragment("(").append(arguments[0]).append(")").
-                            append(strippedOperator).
-                            append("(").append(arguments[2]).append(")");
-                }
-
-
-            }, JdbcType.VARCHAR, 3, 3);
-        }
-
-        QueryService.get().registerPassthroughMethod("to_json", null, JdbcType.OTHER, 1, 1, postgresDialect);
-        QueryService.get().registerPassthroughMethod("to_jsonb", null, JdbcType.OTHER, 1, 1, postgresDialect);
-        QueryService.get().registerPassthroughMethod("array_to_json", null, JdbcType.OTHER, 1, 2, postgresDialect);
-        QueryService.get().registerPassthroughMethod("row_to_json", null, JdbcType.OTHER, 1, 2, postgresDialect);
-
-        addJsonPassthroughMethod("build_array", JdbcType.OTHER, 1, Integer.MAX_VALUE, postgresDialect);
-        addJsonPassthroughMethod("build_object", JdbcType.OTHER, 1, Integer.MAX_VALUE, postgresDialect);
-        addJsonPassthroughMethod("object", JdbcType.OTHER, 1, 2, postgresDialect);
-
-        addJsonPassthroughMethod("array_length", JdbcType.INTEGER, 1, 1, postgresDialect);
-        addJsonPassthroughMethod("each", JdbcType.OTHER, 1, 1, postgresDialect);
-        addJsonPassthroughMethod("each_text", JdbcType.OTHER, 1, 1, postgresDialect);
-        addJsonPassthroughMethod("extract_path", JdbcType.OTHER, 2, Integer.MAX_VALUE, postgresDialect);
-        addJsonPassthroughMethod("extract_path_text", JdbcType.VARCHAR, 2, Integer.MAX_VALUE, postgresDialect);
-        addJsonPassthroughMethod("object_keys", JdbcType.OTHER, 1, 1, postgresDialect);
-        addJsonPassthroughMethod("populate_record", JdbcType.OTHER, 2, 2, postgresDialect);
-        addJsonPassthroughMethod("populate_recordset", JdbcType.OTHER, 2, 2, postgresDialect);
-        addJsonPassthroughMethod("array_elements", JdbcType.OTHER, 1, 1, postgresDialect);
-        addJsonPassthroughMethod("array_elements_text", JdbcType.VARCHAR, 1, 1, postgresDialect);
-        addJsonPassthroughMethod("typeof", JdbcType.VARCHAR, 1, 1, postgresDialect);
-        addJsonPassthroughMethod("to_record", JdbcType.OTHER, 1, 1, postgresDialect);
-        addJsonPassthroughMethod("to_recordset", JdbcType.OTHER, 1, 1, postgresDialect);
-        addJsonPassthroughMethod("strip_nulls", JdbcType.OTHER, 1, 1, postgresDialect);
-
-        QueryService.get().registerPassthroughMethod("jsonb_set", null, JdbcType.OTHER, 3, 4, postgresDialect);
-        QueryService.get().registerPassthroughMethod("jsonb_pretty", null, JdbcType.VARCHAR, 1, 1, postgresDialect);
-    }
-
-    private void addJsonPassthroughMethod(String name, JdbcType type, int minArgs, int maxArgs, SqlDialect dialect)
-    {
-        QueryService.get().registerPassthroughMethod("json_" + name, null, type, minArgs, maxArgs, dialect);
-        QueryService.get().registerPassthroughMethod("jsonb_" + name, null, type, minArgs, maxArgs, dialect);
     }
 
     @Override
